Add files via upload
Forecasting home prices is an important financial decision, not only for those working in
the real estate industry but also for potential home buyers. Because I live and work in Ho
Chi Minh City, I decided to apply the Linear Regression, Random Forest Regression and
Decision Tree Regression models to find the best house price prediction model in Ho Chi
Minh City based on data taken from the ChoTot site.
The ultimate goal of the project is to build a machine learning model to help home buyers
and sellers in Vietnam find a fair price for their homes.